+++ tests/playwright/archive/issue-ac-matrices/strk-121-market-filter-cache-invalidation.spec.js
@@ -25,8 +25,8 @@
  * STAK-521 quarantine predicate inside _loadMarketFilter.
  */
 
-import { test, expect } from "../helpers/mocks/extended-test.js";
-import { injectSeedInventory } from "../helpers/seed.js";
+import { test, expect } from "../../helpers/mocks/extended-test.js";
+import { injectSeedInventory } from "../../helpers/seed.js";
 
 const MARKET_FILTER_KEY = "staktrakr.market_filter";